% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_nba_vars.R
\name{cor_nba_vars}
\alias{cor_nba_vars}
\title{Correlation Matrix of NBA Stats Numerical Variables for a Given Year}
\usage{
cor_nba_vars(yr)
}
\arguments{
\item{yr}{Enter the season.}
}
\description{
This function allows you to find the correlation matrix of the numerical variables in the Kaggle Season_Stats_NBA.csv dataset (found at https://www.kaggle.com/drgilermo/nba-players-stats/data) for any given NBA season.
}
\examples{
cor_nba_vars(2001)
cor_nba_vars()
}
\keyword{Correlation}
\keyword{Matrix}
\keyword{NBA}
