% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mj_game_avg_in_yr.R
\name{mj_game_avg_in_yr}
\alias{mj_game_avg_in_yr}
\title{Michael Jordan Average in Year}
\usage{
mj_game_avg_in_yr(yr)
}
\arguments{
\item{yr}{Enter the season.}
}
\description{
This function allows you to find the average points Michael Jordan scored per game in any given NBA season.
}
\examples{
mj_game_avg_in_yr(1995)
mj_game_avg_in_yr()
}
\keyword{Jordan}
\keyword{Michael}
\keyword{average}
\keyword{points}
