% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player_per_yr.R
\name{oldest_player_per_yr}
\alias{oldest_player_per_yr}
\title{Oldest NBA Player Per Year}
\usage{
oldest_player_per_yr(yr)
}
\arguments{
\item{yr}{Enter the season.}
}
\description{
This function allows you to find the oldest NBA player in a given season.
}
\examples{
oldest_player_per_yr(2015)
oldest_player_per_yr()
}
\keyword{NBA}
\keyword{Oldest}
\keyword{player}
